Add files via upload
The original version of these files originates from a Python Motion Platform Controller for Middlesex University.   The goal is to understand the math behind a stewart platform and then build a servo based stewart platform that can be controlled via a joy stick and canned programs based on motion feedback.